nfs-utils: make substitutions more robust, fix build on musl
replace-fail will complain if there was no occurance of the search
string.

The sed invocation, which was previously used to insert an include at the top of support/nsm/rpc.c has been unnecessary since several releases
but this went unnoticed.

Instead, a new backported includes fix has been added in the form of a patch, so that it will fail to apply once we switch to a release
which includes said patch and we will not forget to remove it.

Change-Id: I9cf589b91dc89135bd7e4e24f85ff04cf953f74b
Signed-off-by: Yureka <yureka@cyberchaos.dev>